This is most of the solaris patch from Jozsef Kadlecsik. Some things I had to change a bit, some things a lot.
- djgpp actually did have mmap, but we don't want to actually use it
  in every possible place.
- The *.init script changes were not acceptable, as they break
  switching between runlevels on both redhat and debian.
- Tested and fixed compilation for NetBSD1.3 and FreeBSD2.2.5.
- Currently djgpp and cygwin are completely broken, the `feature
  detection' is done with the normal compiler and not the correct
  cross-compiler, this will require some more configure changes.

All in all, I went through every line twice, caught some bugs, but
because of the scale of this patch. Something might still break.

+++ coda-src/auth2/ctokens.c
@@ -69,7 +69,7 @@ int main(int argc, char *argv[]) {
 #ifdef __CYGWIN32__
     fprintf(stdout, "Local username: %s\n", getlogin());
 #else
-    fprintf(stdout, "Local uid: %d\n", getuid());
+    fprintf(stdout, "Local uid: %d\n", (int) getuid());
 #endif
 
     /* Get the tokens.  */
